Sort all index files by domain and title
**All index files except** index-domains

When this repository started, the intent was to sort documents in a way where you could click each link and read each thing from the top down, and have each document build on what came before it. As time moved on, the approach shifted towards simply appending documents toward the end of each section. Nowadays, efforts towards beginner-friendly metaphysical resources are being put elsewhere, and this project is now solely oriented towards data preservation.

Because of this, documents in indexes are now sorted by domain and title rather than percieved value. As a result, indexes and sections are now considered categories rather a progressive advancement reading path.

This is also necessary to make it easier to organize the planned document backups, both in the archives repo and in glassport.